Manage app relaunch from click on notification.
Fix app relaunch that was not working in some Android 4.4 and 6 (at least).

Moreover, statistics of opened messages are sent when user clicks on notification and not when message is received.
This because, now, a message can be received on background as well, and this does not imply that it was opened by user.
This behaviour is now aligned with iOS.